ai-agents/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Google Cloud & Vertex AI
google-cloud-aiplatform==1.38.0
google-auth==2.25.2
vertexai==0.0.1

# AI & LLM (Optional - for fallback)
openai==1.3.7
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Vector Store (Optional)
faiss-cpu==1.7.4

# Utilities
python-dateutil==2.8.2